s3_image_croper_ui_app/ImageUiApp/pyproject.toml
[project]
name = "imageuiapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Pi ePaper Project Contributors"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.52.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "ipython (>=9.8.0,<10.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "boto3 (>=1.42.3,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyhere (>=1.0.3,<2.0.0)",
    "mypy (>=1.19.0,<2.0.0)"
]


[dependency-groups]
dev = [
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-cov (>=6.0,<7.0)",
    "ruff (>=0.8,<1.0)",
    "moto (>=5.1.0,<6.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
imageuiapp = "imageuiapp.main:main"
